+++ test/cases/entry-sass-data-url/webpack.config.js
@@ -1,0 +1,53 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+const isProduction = false;
+
+module.exports = {
+  mode: isProduction ? 'production' : 'development',
+  devtool: isProduction ? false : 'source-map',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    styles: './src/assets/scss/main.scss',
+  },
+
+  plugins: [
+    new PugPlugin({
+      modules: [
+        PugPlugin.extractCss({
+          filename: isProduction ? '[name].[contenthash:8].css' : '[name].css',
+        }),
+      ],
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(pug)$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: [
+          {
+            loader: 'css-loader',
+            options: {},
+          },
+          {
+            loader: 'sass-loader',
+            options: {},
+          },
+        ],
+      },
+    ],
+  },
+};